Int -> Int64

+++ src/LN/T/Param.hs
@@ -27,8 +27,8 @@ import           GHC.Generics                (Generic)
 import           Haskell.Api.Helpers.Shared  (QueryParam, qp)
 
 data Param
-  = Limit !(Int)
-  | Offset !(Int)
+  = Limit !(Int64)
+  | Offset !(Int64)
   | SortOrder !(SortOrderBy)
   | Order !(OrderBy)
   | ByOrganizationId !(Int64)
